refactor: collapse installation barrel into installation/index.ts
